Added far too much than that I could remember all the changes. The most important ones are: Implemented first attempt A* pathfinding, allowed entity to move by actually putting the PacketEntityMovement to good use. Entities seem a little too happy, though, as they are dancing around due to their motion not being set accordingly and the default gravition thus never being cleared. Check out the source code for further changes.

+++ gomint-server/src/main/java/io/gomint/server/GoMintServer.java
@@ -168,7 +168,7 @@ public class GoMintServer implements GoMint {
 				this.worldManager.update( currentMillis, lastTickTime );
 
 				long diff = System.nanoTime() - start;
-                lastTickTime = diff / 1000000;
+                lastTickTime = (float) diff / 1000000.0F;
 
 				if ( diff < skipNanos ) {
                     tickCondition.await( skipNanos - diff, TimeUnit.NANOSECONDS );
